requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-google-genai
langchain-community
PyPDF2
python-multipart
python-dotenv
chromadb
sentence-transformers
